refactor: move the shared query-parameter declarations to lib
lib/ already holds the shared request-layer helpers (client_ip,
validation, time); the Annotated query-parameter aliases belong with
them now that the api package is dissolving into domain packages.
