+++ geometrikks/api/v1/analytics_controller.py
@@ -42,7 +42,7 @@ from geometrikks.api.dependencies import (
     provide_live_stats_repo,
     provide_summary_stats_repo
 )
-from geometrikks.api.parameters import (
+from geometrikks.lib.parameters import (
     CityFilter,
     CountryCodeFilter,
     EndTimestamp,
